Just moving somethings around.  As the amount of code for posts and comments increased it started to make more sense for Posts to be in its new place.
