+++ client/src/views/post/index.js
@@ -1,0 +1,2 @@
+export { Comments } from './Comments';
+export { Posts } from './Posts';
