+++ client/src/views/watercooler/Movies.js
@@ -2,7 +2,7 @@ import React, { useContext } from 'react';
 import { UserContext } from 'data/UserStore';
 import MovieCalendar from 'views/MovieCalendar/MovieCalendar';
 import { Interest, Sidebar, Content, Ads } from 'layout';
-import { Posts } from 'views/user';
+import { Posts } from 'views/post';
 
 const Movies = () => {
   const [state, dispatch] = useContext(UserContext);
